move all the JS tests into the test dir
