+++ jest.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  roots: ['<rootDir>/test'],
   setupFiles: ['<rootDir>/jest-setup.js'],
   setupFilesAfterEnv: ['<rootDir>/test/jest-matchers.js'],
   transform: {},
