+++ package.json
@@ -4,8 +4,8 @@
   "description": "Provides access the IPC system calls on unix based platforms",
   "main": "index.js",
   "scripts": {
-    "test": "node node_modules/jest/bin/jest.js --roots src",
-    "test:inspect": "node --inspect-brk node_modules/jest/bin/jest.js --runInBand --testTimeout=60000 --collectCoverage=false --roots src",
+    "test": "node node_modules/jest/bin/jest.js",
+    "test:inspect": "node --inspect-brk node_modules/jest/bin/jest.js --runInBand --testTimeout=60000 --collectCoverage=false",
     "build:debug": "swig -c++ -o src/OSX.cpp -javascript -napi src/OSX.i && node-gyp rebuild --debug",
     "prepare": "husky",
     "install": "swig -c++ -o src/OSX.cpp -javascript -node src/OSX.i && node-gyp rebuild",
