+++ test/semaphore-posix.test.js
@@ -1,5 +1,5 @@
 const { fork } = require('node:child_process');
-const childMessages = require('../test/parent.js');
+const childMessages = require('./parent.js');
 const { SemaphoreP } = require('../build/Release/OSX.node');
 
 const name = Buffer.from('semaphore-test').toString('base64');
